Isolated stepper into it's own component separate from the NewProjectForm view. Ready to start creating the form inputs.
